fix(service): :bug: assign listener from paginator when closeable enabled
The closeable was already assigned to a listener when there were more than one page provided to the paginator, but not when it was only one page. Now it is added to the non-pagination close listener if there is one page, as if it were the normal kind of closeable.