feat(service): :boom: add starboard feature & service changes
Added the full starboard feature (classic in discord bots). The feature is a service with an accompanying command. The service architecture was also changed.

Service properties and execution were changed dramatically. Services must now specify the events it should be executed for, and will receive the exact args related to that event (must prepare for type checking). In the event listeners, any services related to that event will be executed.