fix(service): :bug: resolve shotty mute evasion countervention attempt
Instead of checking on all unmute timedactions and trying to find the right one, which would only work if the user was muted for a certain amount of time, just check if the user had the server's muted role when they left. There is a separate role cache now that is updated when a user leaves. This can also be used for another autorole feature in the future.