fix(command): :art: better nickname perm check and fallback message + code quality
Brings to the current code quality standard of this repo and checks to see whether the target is the owner (to make a more specific fallback message). Additionally: fixed a bug where if the nickname (after specifying the sender as the target) began with the id of the sender, that part of the nickname would be removed.