fix(service): :bug: starboard no post bug + emojis bug + full button support
Fixed the bug where sometimes messages wouldn't be seen and poted in their channels. Other issue was that the emoji flag and emoji whatever wouldn't actually take emojis like that, so support had to be added for them in the regex. The final feature is full use of the buttons.