fix(core): :arrow_down: add nodemon for unidentified dev runtime error (windows-kill.exe)
An error occurred (seemingly randomly) during development saying that the windows-kill.exe module could not be found. This seems to be a necessary dependency for nodemon to have in order to kill (and restart) windows node processes. It couldn't find it, I don't know why, I don't remember messing with dependencies.