fix(core): :bug: move flag parsing ahead of cmd args length check
This was done because for commands that require arguments (not flags, normal arguments), the script would obviously stop execution before the flags were parsed if the user did not provide the necessary number of arguments. Of course, the number of arguments would also be inaccurate because it would count the number of flags at the same time. The overall purpose of this was so the --help argument could be seen before the script stopped.